Corrigindo teste da api
