+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start -p $PORT",
     "lint": "next lint",
-    "test": "jest --coverage --maxWorkers 2"
+    "test": "jest --coverage --maxWorkers 5"
   },
   "dependencies": {
     "@chakra-ui/react": "^2.2.6",
